helios: per-fence name discriminator in present sync
One producer process owns several D3D11 devices (dwm creates at least
two — proven live: the second device's named-fence creation failed on
the per-pid name collision and disabled its publish path). Names are
now Global\HeliosPresentFence_<pid>_<fenceId> with a process-unique
fence id carried in the publication slot; the consumer caches imports
by (pid, fenceId), so a recreated device invalidates naturally.

+++ src/dxvk/dxvk_context.h
@@ -1438,13 +1438,14 @@ namespace dxvk {
     uint32_t                                  m_heliosStagedProbesIssued = 0u;
 
     /* Helios WS1 #4 consumer-side present wait (dxvk.heliosPresentWaitUs).
-     * Named present fences imported once per producer pid; negative cache
-     * (null fence) for unresolvable names with periodic retry. */
+     * Named present fences imported once per (producer pid << 32 | fence
+     * id); negative cache (null fence) for unresolvable names with periodic
+     * retry. */
     struct HeliosPresentWaitFence {
       Rc<DxvkFence> fence;
       uint32_t      retryCountdown = 0u;
     };
-    std::unordered_map<uint32_t, HeliosPresentWaitFence> m_heliosPresentWaitFences;
+    std::unordered_map<uint64_t, HeliosPresentWaitFence> m_heliosPresentWaitFences;
     uint64_t                                  m_heliosPresentWaits        = 0u;
     uint64_t                                  m_heliosPresentWaitUsTotal  = 0u;
     uint64_t                                  m_heliosPresentWaitTimeouts = 0u;
